// src/types/slices/task.ts
export type Task = {
    id: number
    idspecific_objective: number
    idcommittee: number
    committee: string
    idrrhh: number
    rrhh: string
    n_document_rrhh: string
    idcommittee_state: number
    state: string
    code: string
    name: string
    description: string
    start_date: string
    end_date: string
    status: string
}